+++ tom_demark_indicator/__init__.py
@@ -3,6 +3,7 @@ from .data_loader import load_data
 from .indicators import add_indicators
 from .td_sequential import add_td_sequential
 from .plotting_mpf import plot_with_mplfinance
+from .exporter import save_data_json, default_image_path
 
 __all__ = [
     "PlotConfig",
@@ -10,4 +11,6 @@ __all__ = [
     "add_indicators",
     "add_td_sequential",
     "plot_with_mplfinance",
+    "save_data_json",
+    "default_image_path",
 ]